trying to remove some warnings from display.cpp when running on VM

+++ src/display.cpp
@@ -12,7 +12,7 @@ namespace dsply {
             for (size_t i = 0; i < piece->orientations.size(); ++i) {
                 displayPiece(piece, i);
             }
-        } else if (orientation_idx < -1 || orientation_idx >= piece->orientations.size()) {
+        } else if (orientation_idx < -1 || orientation_idx >= (int)piece->orientations.size()) {
             std::cerr << msgs::outOfBoundsOrientationIndexDisplayPieceErrorMessage << std::endl;
             return;
         } else {
@@ -40,7 +40,7 @@ namespace dsply {
             displayStructure.push_back(currentVec);
 
             //Display
-            for (int i = 0; i < displayStructure.size(); ++i) {
+            for (size_t i = 0; i < displayStructure.size(); ++i) {
                 //invariant: colTracker is where we expect to place the next char in the current row if it's as far left as possible, given the previously placed pieces in this row (starting from firstCol)
                 int colTracker = firstCol;
                 //j starts at 1 because the first element in the vector is reserved for the spatial row
